akher edit

+++ database/migrations/2023_05_06_124025_create_wishlists_table.php
@@ -14,14 +14,14 @@ return new class extends Migration
     {
         Schema::create('wishlists', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('product_id');
+            $table->integer('user_id');
+            $table->integer('product_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
 
-           
-            
+
+
         });
     }
